adding filtres in utils
